+++ src/main/scala/com/yayetee/winko/AppLauncher.scala
@@ -40,10 +40,10 @@ object AppLauncher extends Application {
 
 	}
 	
-	def createSymbol(symbolID: Int, xpos: Float, ypos: Float, angle: Float) = {
+	def createSymbol(symbolID: Int, xpos: Double, ypos: Double, angle: Double) = {
 		Mash.logger.debug("createSymbol appLauncher")		
 		new MashSymbol(xpos, ypos, angle)
 	}
 
-	def createCursor(xpos: Float, ypos: Float) = new MashCursor(xpos, ypos)
+	def createCursor(xpos: Double, ypos: Double) = new MashCursor(xpos, ypos)
 }
